requirements.txt
streamlit==1.38.0
sqlmodel==0.0.21
sqlalchemy==2.0.35
pydantic==1.10.15
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pandas==2.2.2
openpyxl==3.1.5
XlsxWriter==3.2.0
python-dateutil==2.9.0.post0
